Add tests about already present parachains  weights
