+++ runtime/src/tests/use_correct_weights.rs
@@ -1,6 +1,23 @@
+// Copyright 2024, Horizen Labs, Inc.
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+
 //! Here we write the integration tests that just check pallets weighs are correctly linked.
 
 use crate::{OnChainSeqPhragmen, Runtime};
+
+mod verifiers;
 
 #[test]
 fn frame_system() {
@@ -201,70 +218,6 @@ fn pallet_im_online() {
 }
 
 #[test]
-fn pallet_fflonk_verifier() {
-    use pallet_fflonk_verifier::Fflonk;
-    let dummy_proof = [0; pallet_fflonk_verifier::PROOF_SIZE];
-    let dummy_pubs = [0; pallet_fflonk_verifier::PUBS_SIZE];
-    use pallet_fflonk_verifier::WeightInfo;
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Fflonk>>::WeightInfo as pallet_verifiers::WeightInfo<Fflonk>>::submit_proof(
-            &dummy_proof,
-            &dummy_pubs
-        ),
-        crate::weights::pallet_fflonk_verifier::ZKVWeight::<Runtime>::submit_proof()
-    );
-}
-
-#[test]
-fn pallet_zksync_verifier() {
-    use pallet_zksync_verifier::Zksync;
-    let dummy_proof = [0; pallet_zksync_verifier::PROOF_SIZE];
-    let dummy_pubs = [0; pallet_zksync_verifier::PUBS_SIZE];
-    use pallet_zksync_verifier::WeightInfo;
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Zksync>>::WeightInfo as pallet_verifiers::WeightInfo<Zksync>>::submit_proof(
-            &dummy_proof,
-            &dummy_pubs
-        ),
-        crate::weights::pallet_zksync_verifier::ZKVWeight::<Runtime>::submit_proof()
-    );
-}
-
-#[test]
-fn pallet_groth16_verifier() {
-    use pallet_groth16_verifier::Groth16;
-    use pallet_groth16_verifier::WeightInfo;
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Groth16<Runtime>>>::WeightInfo as
-            pallet_verifiers::WeightInfo<Groth16<Runtime>>>
-            ::submit_proof(
-            &pallet_groth16_verifier::Proof::default(),
-            &Vec::new()
-        ),
-        crate::weights::pallet_groth16_verifier::ZKVWeight::<Runtime>::submit_proof_bn254(0)
-    );
-}
-
-#[test]
-fn pallet_settlement_risc0() {
-    use pallet_risc0_verifier::Risc0;
-    use pallet_risc0_verifier::WeightInfo;
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Risc0<Runtime>>>::WeightInfo as
-            pallet_verifiers::WeightInfo<Risc0<Runtime>>>
-            ::submit_proof(
-            &Vec::new(),
-            &Vec::new()
-        ),
-        crate::weights::pallet_risc0_verifier::ZKVWeight::<Runtime>::submit_proof_cycle_2_pow_13()
-    );
-}
-
-#[test]
 fn pallet_settlement_ultraplonk() {
     use pallet_ultraplonk_verifier::{Ultraplonk, WeightInfo};
 
@@ -297,4 +250,72 @@ fn pallet_bags_list() {
         <Runtime as pallet_bags_list::Config<pallet_bags_list::Instance1>>::WeightInfo::put_in_front_of(),
         crate::weights::pallet_bags_list::ZKVWeight::<Runtime>::put_in_front_of()
     );
+}
+
+mod parachains {
+    #![cfg(feature = "relay")]
+
+    use super::*;
+    use crate::parachains::*;
+
+    #[test]
+    fn configuration() {
+        use configuration::WeightInfo;
+
+        assert_eq!(
+            <<Runtime as configuration::Config>::WeightInfo as
+            configuration::WeightInfo>::set_config_with_block_number(),
+            crate::weights::parachains::configuration::ZKVWeight::<Runtime>::set_config_with_block_number()
+        )
+    }
+
+    #[test]
+    fn disputes() {
+        use disputes::WeightInfo;
+
+        assert_eq!(
+            <<Runtime as disputes::Config>::WeightInfo as disputes::WeightInfo>::force_unfreeze(),
+            crate::weights::parachains::disputes::ZKVWeight::<Runtime>::force_unfreeze()
+        )
+    }
+
+    #[test]
+    fn hrmp() {
+        use hrmp::WeightInfo;
+
+        assert_eq!(
+            <<Runtime as hrmp::Config>::WeightInfo as hrmp::WeightInfo>::hrmp_init_open_channel(),
+            crate::weights::parachains::hrmp::ZKVWeight::<Runtime>::hrmp_init_open_channel()
+        )
+    }
+
+    #[test]
+    fn initializer() {
+        use initializer::WeightInfo;
+
+        assert_eq!(
+            <<Runtime as initializer::Config>::WeightInfo as initializer::WeightInfo>::force_approve(42),
+            crate::weights::parachains::initializer::ZKVWeight::<Runtime>::force_approve(42)
+        )
+    }
+
+    #[test]
+    fn paras_inherent() {
+        use paras_inherent::WeightInfo;
+
+        assert_eq!(
+            <<Runtime as paras_inherent::Config>::WeightInfo as paras_inherent::WeightInfo>::enter_bitfields(),
+            crate::weights::parachains::paras_inherent::ZKVWeight::<Runtime>::enter_bitfields()
+        )
+    }
+
+    #[test]
+    fn paras() {
+        use paras::WeightInfo;
+
+        assert_eq!(
+            <<Runtime as paras::Config>::WeightInfo as paras::WeightInfo>::force_set_most_recent_context(),
+            crate::weights::parachains::paras::ZKVWeight::<Runtime>::force_set_most_recent_context()
+        )
+    }
 }
